requirements.txt
asgiref==3.10.0
coverage==7.11.3
Django==3.0
entrypoints==0.3
factory_boy==3.3.3
Faker==37.12.0
flake8==3.7.0
iniconfig==2.3.0
isort==7.0.0
mccabe==0.6.1
packaging==25.0
pluggy==1.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.19.2
pytest==8.4.2
pytest-django==3.9.0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
